2021.07.07 mysite03 guestbook jQuery, Ajax practice

+++ mysite03/src/main/java/com/douzone/mysite/controller/GuestbookController.java
@@ -34,7 +34,7 @@ public class GuestbookController {
 	}
 	
 	@RequestMapping(value="/delete/{no}", method=RequestMethod.POST)
-	public String delete(@PathVariable("no") Long no, @RequestParam(value="password", required=true, defaultValue="") String password) {
+	public String delete(@PathVariable("no") int no, @RequestParam(value="password", required=true, defaultValue="") String password) {
 		// request.setCharacterEncoding("utf-8");
 		//guestbookService.delete(no, password);
 		guestbookService.deleteMessage(no, password);
